Initial public version of the Situm SDK for JS.
Please bare in mind this is the first draft version. You must not use it as a
production-ready package. We will publish further versions that stability the
module.

This version includes the first consummer wrappers for the SITUM REST API for
buildings and their cartography (floors, geofences, pois, paths, events),
users and fetching realtime positioning. It was added multiple authentication methods.

Ref: PRO-1676